cambio en la columna Amount de las apuestas de usuario y script de inicio con los deportes

+++ ADOL.APP/ADOL.APP.DataBase/dbo/Tables/ApuestasDeUsuarios.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[ApuestasDeUsuarios] (
     [ID]                 INT          IDENTITY (1, 1) NOT NULL,
     [Token]              VARCHAR (36) NOT NULL,
-    [Amount]             DECIMAL (18) NOT NULL,
+    [Amount]             FLOAT NOT NULL,
     [ApuestaDeportivaID] INT          NOT NULL,
     [Acierto]            BIT          NULL,
     CONSTRAINT [PK_ApuestasDeUsuarios] PRIMARY KEY CLUSTERED ([ID] ASC),
